Differentiate through preconditioner logdet for variance reduction
